+++ src/drivers/main.cpp
@@ -4,6 +4,7 @@
 #include "precice/config/Configuration.hpp"
 #include <iostream>
 #include "logging/Logger.hpp"
+#include "xml/Printer.hpp"
 
 void printUsage()
 {
@@ -11,6 +12,7 @@ void printUsage()
   std::cout << "Run server (deprecated)  :  binprecice server ParticipantName ConfigurationName [LogConfFile]\n";
   std::cout << "Print XML reference      :  binprecice xml\n";
   std::cout << "Print DTD for XML config :  binprecice dtd" << std::endl;
+  std::cout << "Print Markdown reference :  binprecice md" << std::endl;
 }
 
 int main ( int argc, char** argv )
@@ -18,6 +20,7 @@ int main ( int argc, char** argv )
   bool runServer = false;
   bool runHelp = false;
   bool runDtd = false;
+  bool runMD = false;
   bool hasLogConfFile = false;
 
   bool wrongParameters = true;
@@ -27,6 +30,10 @@ int main ( int argc, char** argv )
     if ( action == "dtd" ) {
       wrongParameters = false;
       runDtd = true;
+    }
+    if ( action == "md" ) {
+      wrongParameters = false;
+      runMD = true;
     }
     if ( action == "xml" ) {
       wrongParameters = false;
@@ -76,12 +83,17 @@ int main ( int argc, char** argv )
   else if (runHelp){
     PRECICE_ASSERT(not runServer);
     precice::config::Configuration config;
-    std::cout << config.getXMLTag().printDocumentation(0) << "\n\n";
+    precice::xml::toDocumentation(std::cout, config.getXMLTag());
   }
   else if (runDtd) {
 	PRECICE_ASSERT(not runServer);
     precice::config::Configuration config;
-    std::cout << config.getXMLTag().printDTD(true) << "\n\n";
+    precice::xml::toDTD(std::cout, config.getXMLTag());
+  }
+  else if (runMD) {
+	PRECICE_ASSERT(not runServer);
+    precice::config::Configuration config;
+    precice::xml::toMarkdown(std::cout, config.getXMLTag());
   }
   else {
     PRECICE_ASSERT( false );